Move useful spec helpers into SpecContextHelpers.scala.
We added a few spec helpers to `LiftMergeSpec`, but these helpers were
more broadly useful, so we now move them to a common file.
